pyproject.toml
[tool.poetry]
name = "yoshimidi"
version = "0.1.0"
description = ""
authors = ["Misha Wagner <mishajw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fire = "^0.5.0"
requests = "^2.31.0"
tqdm = "^4.65.0"
msgspec = "^0.16.0"
loguru = "^0.7.0"
pygame = "^2.5.0"
jaxtyping = "^0.2.20"
seaborn = "^0.12.2"
mido = "^1.2.10"
torch = "2.0.0"
pytest = "^7.4.0"
wandb = "^0.15.5"
python-dotenv = "^1.0.0"
jsonlines = "^3.1.0"
pydantic = "^2.0.3"
toml = "^0.10.2"

[tool.poetry.group.dev]
optional = true
[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
mypy = "^1.4.0"
types-requests = "^2.31.0.1"
ipykernel = "^6.23.3"
types-tqdm = "^4.65.0.1"
types-toml = "^0.10.8.6"
vastai = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
allow_redefinition = true
plugins = ["pydantic.mypy"]
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true
# TODO: Resolve issues with numpy generics.
# disallow_any_generics = true

[[tool.mypy.overrides]]
module = ["mido.*", "fire"]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.isort]
profile = "black"
